adding url driven data base access 4 (edit db records)

+++ php/removeRow.php
@@ -1,32 +1,35 @@
 <?php
-include 'db_path.php';
+//include 'db_path.php';
 
-// get user IP address
-$ip = $_SERVER['REMOTE_ADDR'];
+if (isset($_POST['path']) && isset($_POST['rmr'])) {
+   $variable_to_share = $_POST['path'];
+   $raw_post_string = $_POST['rmr'];
+   // get user IP address
+   $ip = $_SERVER['REMOTE_ADDR'];
 
-//gets chart info from web page
-$raw_post_string = file_get_contents('php://input');
+   //gets chart info from web page
+   //$raw_post_string = file_get_contents('php://input');
 
-// decodes info from web pages
-$array = json_decode($raw_post_string, true);
+   // decodes info from web pages
+   $array = json_decode($raw_post_string, true);
 
-// makes a connection to the database
-$db = new PDO('sqlite:'.$variable_to_share);
+   // makes a connection to the database
+   $db = new PDO('sqlite:'.$variable_to_share);
 
-if ($db != null)
-    echo " db connected ";
-else
-    echo " db did not connect ";
+    if ($db != null)
+        echo " db connected ";
+    else
+        echo " db did not connect ";
 
-// makes a variable that is the length of the array that is the webpage info.
-var_dump($array[0]);
+    // makes a variable that is the length of the array that is the webpage info.
+    var_dump($array[0]);
 
-// creates an database input string
-$stm = $db->prepare("delete from eventTable where plotId=:plotid and User_IP=:userid");
+    // creates an database input string
+    $stm = $db->prepare("delete from eventTable where plotId=:plotid and User_IP=:userid");
 
-$stm->bindValue(':plotid',$array[0]);
-$stm->bindValue(':userid',$array[1]);
-
+    $stm->bindValue(':plotid',$array[0]);
+    $stm->bindValue(':userid',$array[1]);
+}
 // executes database input string with above varaibles
 if ($stm->execute()) { 
 	// it worked
